# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"IUser": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female"
						]
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"firstName",
					"lastName",
					"gender",
					"DOB",
					"verified",
					"status",
					"accountType",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISession": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string",
						"enum": [
							"group",
							"individual"
						]
					},
					"serviceProviderID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IServiceProvider"
							}
						]
					},
					"clientsID": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClient"
								},
								"type": "array"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"online",
							"home",
							"office"
						]
					},
					"location": {
						"additionalProperties": false,
						"type": "object"
					},
					"attachments": {
						"items": {
							"properties": {
								"type": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"type",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"requirements": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"properties": {
								"ratingDate": {
									"type": "string",
									"format": "date-time"
								},
								"ratingValue": {
									"type": "string"
								},
								"raterUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								}
							},
							"required": [
								"ratingDate",
								"ratingValue",
								"raterUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"properties": {
								"reviewDate": {
									"type": "string",
									"format": "date-time"
								},
								"reviewDetails": {
									"type": "string"
								},
								"reviewerUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								}
							},
							"required": [
								"reviewDate",
								"reviewDetails",
								"reviewerUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"properties": {
								"amount": {
									"type": "number",
									"format": "double"
								},
								"payerID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								},
								"paymentMethod": {
									"type": "string"
								},
								"discount": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"amount",
								"payerID",
								"paymentMethod",
								"discount"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"finished"
						]
					},
					"doctorReferral": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"sessionType",
					"name",
					"serviceType",
					"location",
					"requirements",
					"sessionFee",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAgreement": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"parties",
					"startDate",
					"endDate",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRole": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"service_provider": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"clients": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sessions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"communications": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"disputes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"enum_values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"employees",
					"users",
					"clients",
					"sessions",
					"communications",
					"disputes",
					"enum_values"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEmployee": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"roleID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IRole"
							}
						]
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"properties": {
								"empType": {
									"type": "string"
								},
								"url": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"empType",
								"url",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IEmployee_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"roleID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IRole"
							}
						]
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"properties": {
								"empType": {
									"type": "string"
								},
								"url": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"empType",
								"url",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"INotification": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Agreement",
							"Dispute",
							"Alarm",
							"Announcement",
							"User",
							"Communication"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/IAgreement"
							},
							{
								"$ref": "#/components/schemas/IDispute"
							},
							{
								"$ref": "#/components/schemas/IAlarm"
							},
							{
								"$ref": "#/components/schemas/IAnnouncement"
							},
							{
								"$ref": "#/components/schemas/IUser"
							},
							{
								"$ref": "#/components/schemas/ICommunication"
							}
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"delivered",
							"opened"
						]
					},
					"title": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"receiverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					}
				},
				"required": [
					"referenceType",
					"referenceID",
					"status",
					"title",
					"details",
					"sentDate",
					"receiverUID"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_INotification_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Agreement",
							"Dispute",
							"Alarm",
							"Announcement",
							"User",
							"Communication"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/IAgreement"
							},
							{
								"$ref": "#/components/schemas/IDispute"
							},
							{
								"$ref": "#/components/schemas/IAlarm"
							},
							{
								"$ref": "#/components/schemas/IAnnouncement"
							},
							{
								"$ref": "#/components/schemas/IUser"
							},
							{
								"$ref": "#/components/schemas/ICommunication"
							}
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"delivered",
							"opened"
						]
					},
					"title": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"receiverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IEnumValues": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"values",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IEnumValues_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"note": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/employees/{employeeId}": {
			"get": {
				"operationId": "GetEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployee"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested employee is not found"
					}
				},
				"description": "Get employee details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				]
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployee"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Delete employee",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				]
			}
		},
		"/employees/create": {
			"post": {
				"operationId": "CreateEmployee",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEmployee"
								},
								"examples": {
									"Example 1": {
										"value": {
											"uID": "63130ba28d796320172e58d6",
											"roleID": "631312818d796320172e58e5",
											"salary": 5000.5,
											"attachments": [
												{
													"name": "memo",
													"url": "www.google.com",
													"empType": "some type"
												}
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed"
					}
				},
				"description": "Create employee",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmployee"
							}
						}
					}
				}
			}
		},
		"/employees/update/{employeeId}": {
			"put": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployee"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update employee",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IEmployee_"
							},
							"example": "6313151791b8bbcc188d8e37"
						}
					}
				}
			}
		},
...
		"/notifications/{notificationId}": {
			"get": {
				"operationId": "GetNotification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/INotification"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested notification in not found"
					}
				},
				"description": "Get a Notification details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteNotification",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/INotification"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested notification in not found"
					}
				},
				"description": "Delete a notification",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313410084cfa62399551059"
					}
				]
			}
		},
		"/notifications/create": {
			"post": {
				"operationId": "CreateNotification",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/INotification"
								},
								"examples": {
									"Example 1": {
										"value": {
											"referenceType": "Session",
											"referenceID": "63131c361ea32f520cb28bd6",
											"status": "opened",
											"title": "Warning",
											"details": "this notification contain on new details",
											"sentDate": "2022-09-10T00:00:00.000Z",
											"receivedDate": "2022-09-10T00:00:00.000Z",
											"openDate": "2022-09-10T00:00:00.000Z",
											"receiverUID": "63130ba28d796320172e58d6"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a notification",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotification"
							}
						}
					}
				}
			}
		},
		"/notifications/update/{notificationId}": {
			"put": {
				"operationId": "UpdateNotification",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/INotification"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a notification",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_INotification_"
							}
						}
					}
				}
			}
		},
...
		"/enums/{enumId}": {
			"get": {
				"operationId": "GetEnum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEnumValues"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested enum in not found"
					}
				},
				"description": "Get a Enums details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "enumId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631370d58ec48bbce3f18fb6"
					}
				]
			},
			"delete": {
				"operationId": "DeleteEnum",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IEnumValues"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested enum in not found"
					}
				},
				"description": "Delete a enum",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "enumId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631370d58ec48bbce3f18fb6"
					}
				]
			}
		},
		"/enums/create": {
			"post": {
				"operationId": "CreateEnum",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEnumValues"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "session",
											"values": [
												"first",
												"second"
											],
											"note": "test"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a enum",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEnumValues"
							}
						}
					}
				}
			}
		},
		"/enums/update/{enumId}": {
			"put": {
				"operationId": "UpdateEnum",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEnumValues"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a enum",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "enumId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631370d58ec48bbce3f18fb6"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IEnumValues_"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}